Various PR fixups Swap assert functions to take an id instead of an object Turn createAdminPanelDBFilter into a util Add a check before pushing demo land permissions when filtering creating an admin panel filter
