+++ packages/meditrak-server/src/routes/userAccounts/assertUserAccountPermissions.js
@@ -8,7 +8,8 @@ import { hasBESAdminAccess, TUPAIA_ADMIN_PANEL_PERMISSION_GROUP } from '../../pe
 
 const { RAW } = QUERY_CONJUNCTIONS;
 
-export const assertUserAccountPermissions = async (accessPolicy, models, userAccount) => {
+export const assertUserAccountPermissions = async (accessPolicy, models, userAccountId) => {
+  const userAccount = await models.user.findById(userAccountId);
   const entityPermissions = await models.userEntityPermission.find({ user_id: userAccount.id });
   const entities = await models.entity.findManyById(entityPermissions.map(ep => ep.entity_id));
   const countryCodes = entities.map(e => e.country_code).filter(c => c !== 'DL');
